pyproject.toml
[project]
name = "nonebot-plugin-ehentai-search"
version = "0.1.4"
authors = [
  { name="N791", email="a15345506127@gmail.com" },
]
description = "一个基于Nonebot2和OnebotV11的ehentai搜索"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}

[tool.nonebot]
plugins = ["nonebot_plugin_ehentai_search"]

[tool.poetry]
name = "nonebot-plugin-ehentai-search"
version = "0.1.4"
description = "ehentai search plugin for NoneBot2"
authors = ["N791"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/N791/nonebot-plugin-ehentai-search"
documentation = "https://github.com/N791/nonebot-plugin-ehentai-search/blob/main/README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
nonebot2 = ">=2.2.0"
nonebot-adapter-onebot = "^2.4.4"
Pillow = "^9.4.0"
beautifulsoup4 = "^4.11.2"
httpx = "^0.23.0"
playwright = "^1.28.1"

[project.urls]
"Homepage" = "https://github.com/N791/nonebot-plugin-ehentai-search"
"Bug Tracker" = "https://github.com/N791/nonebot-plugin-ehentai-search/issues"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"